+++ src/scriv/__init__.py
@@ -2,4 +2,4 @@
 Scriv changelog management tool.
 """
 
-__version__ = "0.8.1"
+__version__ = "0.9.0"